[Fix/#17] MVVM 적용 해제 및 Model 분리
